Wip/paraphrase

+++ paraphrase-job.sh
@@ -3,7 +3,8 @@
 . /opt/genie-toolkit/lib.sh
 
 parse_args "$0" "s3_bucket owner dataset_owner project experiment \
-            input_dataset output_dataset filtering_model paraphrasing_model_full_path skip_generation skip_filtering task_name ignore_context" "$@"
+            input_dataset output_dataset filtering_model paraphrasing_model_full_path \
+            skip_generation skip_filtering keep_original_duplicates task_name ignore_context" "$@"
 shift $n
 
 if [ "$ignore_context" = true ] ; then
@@ -27,7 +28,7 @@ if [ "$task_name" = "almond_dialogue_nlu" ] ; then
   export input_dir="input_dataset/user"
   export output_dir="output_dataset/user"
   export filtering_dir="almond/user"
-  export filtering_batch_size="300"
+  export filtering_batch_size="128"
 elif [ "$task_name" = "almond" ] ; then
   export input_dir="input_dataset"
   export output_dir="output_dataset"
@@ -38,13 +39,15 @@ else
 fi
 
 make_input_ready(){
-    # remove duplicates before paraphrasing to avoid wasting effort
-    python3 /opt/genienlp/genienlp/paraphrase/scripts/transform_dataset.py \
-      ${input_dir}/train.tsv \
-      ${input_dir}/train_no_duplicate.tsv \
-      --remove_duplicates \
-      --task ${task_name}
-      cp ${input_dir}/train_no_duplicate.tsv ${input_dir}/train.tsv
+    if [ "$keep_original_duplicates" = false ] ; then
+      # remove duplicates before paraphrasing to avoid wasting effort
+      python3 /opt/genienlp/genienlp/paraphrase/scripts/transform_dataset.py \
+        ${input_dir}/train.tsv \
+        ${input_dir}/train_no_duplicate.tsv \
+        --remove_duplicates \
+        --task ${task_name}
+        cp ${input_dir}/train_no_duplicate.tsv ${input_dir}/train.tsv
+    fi
 
     if [ "$is_dialogue" = true ] ; then
     # add previous agent utterance; this additional context helps the paraphraser
@@ -63,6 +66,7 @@ run_paraphrase(){
   if [ "$is_dialogue" = true ] ; then
     if [ "$ignore_context" = true ] ; then
       genienlp run-paraphrase \
+      --task paraphrase \
       --model_name_or_path paraphraser \
       --input_file ${output_dir}/with_context.tsv \
       --output_file ${output_dir}/paraphrased.tsv \
@@ -72,6 +76,7 @@ run_paraphrase(){
       $paraphrasing_arguments
     else
       genienlp run-paraphrase \
+      --task paraphrase \
       --model_name_or_path paraphraser \
       --input_file ${output_dir}/with_context.tsv \
       --output_file ${output_dir}/paraphrased.tsv \
@@ -83,6 +88,7 @@ run_paraphrase(){
     fi
   else
     genienlp run-paraphrase \
+      --task paraphrase \
       --model_name_or_path paraphraser \
       --input_file ${input_dir}/train.tsv \
       --output_file ${output_dir}/paraphrased.tsv \
@@ -101,7 +107,6 @@ join(){
     --query_file ${output_dir}/paraphrased.tsv \
     --num_new_queries ${num_new_queries} \
     --transformation replace_queries \
-    --remove_duplicates \
     --remove_with_heuristics \
     --task ${task_name}
 }
@@ -129,22 +134,23 @@ filter(){
   python3 /opt/genienlp/genienlp/paraphrase/scripts/transform_dataset.py \
     ${output_dir}/unfiltered.tsv \
     ${output_dir}/filtered.tsv \
+    --thrown_away ${output_dir}/thrown_away.tsv \
     --thingtalk_gold_file eval_dir/valid/${task_name}.tsv \
     --transformation remove_wrong_thingtalk \
-    --remove_duplicates \
     --task ${task_name}
 }
 
 append_to_original(){
   # append paraphrases to the end of the original training file and remove duplicates
-  cp ${input_dir}/train.tsv ${output_dir}/temp.tsv
-  cat ${output_dir}/filtered.tsv >> ${output_dir}/temp.tsv
+  cp ${input_dir}/train.tsv ${output_dir}/train.tsv
+  cp ${output_dir}/filtered.tsv ${output_dir}/temp.tsv
   python3 /opt/genienlp/genienlp/paraphrase/scripts/transform_dataset.py \
     ${output_dir}/temp.tsv \
-    ${output_dir}/train.tsv \
+    ${output_dir}/filtered.tsv \
     --remove_duplicates \
     --task ${task_name}
   rm ${output_dir}/temp.tsv
+  cat ${output_dir}/filtered.tsv >> ${output_dir}/train.tsv
 }
 
 
@@ -160,6 +166,12 @@ else
 
   # paraphrasing was successful, so upload the intermediate files
   aws s3 sync output_dataset s3://${s3_bucket}/${dataset_owner}/dataset/${project}/${experiment}/${output_dataset}
+  echo "Uploaded intermediate files to s3."
+  if test -f output_dataset/synthetic.txt ; then
+    echo "Deleting synthetic.txt from pod to save disk space..."
+    rm output_dataset/synthetic.txt
+  fi
+  
 fi
 
 if [ "$skip_filtering" = true ] ; then
